refactoring code related to authentication
